aulas/05-aula/requirements.txt
flask==0.10.1 
pyserial==3.0.1 
